Relocate and update capacitor config
