+++ client/packages/android/app/src/main/assets/capacitor.config.json
@@ -1,9 +1,16 @@
 {
 	"appId": "org.openmsupply.client",
 	"appName": "openmsupply-client",
-	"webDir": "packages/host/dist",
+	"webDir": "notexistent",
 	"bundledWebRuntime": false,
 	"android": {
-		"path": "packages/android"
+		"path": "./"
+	},
+	"server": {
+		"url": "https://localhost:8000",
+		"hostname": "should.notmatch.localhost",
+		"allowNavigation": [
+			"*"
+		]
 	}
 }
